Move Supervisor to Application module
