+++ mix.exs
@@ -38,8 +38,8 @@ defmodule Xcribe.MixProject do
     ]
   end
 
-  defp application_mod(:test), do: {Xcribe, children: [{Xcribe.Endpoint, []}]}
-  defp application_mod(_), do: {Xcribe, []}
+  defp application_mod(:test), do: {Xcribe.Application, children: [{Xcribe.Endpoint, []}]}
+  defp application_mod(_), do: {Xcribe.Application, []}
 
   defp deps do
     [
